+++ lib/datamappify/data/criteria/active_record/find.rb
@@ -1,5 +1,3 @@
-require 'datamappify/data/criteria/relational/find'
-
 module Datamappify
   module Data
     module Criteria