+++ lib/datamappify/data/criteria/active_record/limit.rb
@@ -1,8 +1,8 @@
 module Datamappify
   module Data
     module Criteria
-      module Sequel
-        class Find < Relational::Find
+      module ActiveRecord
+        class Limit < Relational::Limit
         end
       end
     end