+++ lib/datamappify/data/criteria/active_record/save_by_key.rb
@@ -1,4 +1,3 @@
-require 'datamappify/data/criteria/relational/save_by_key'
 require 'datamappify/data/criteria/active_record/save'
 
 module Datamappify