+++ lib/datamappify/data/criteria/sequel/count.rb
@@ -1,5 +1,3 @@
-require 'datamappify/data/criteria/relational/count'
-
 module Datamappify
   module Data
     module Criteria