+++ lib/datamappify/repository/query_method/criteria.rb
@@ -4,7 +4,9 @@ module Datamappify
       class Criteria < Method
         # @return [Array<Entity>]
         def perform
-          dispatch_criteria_to_default_source(:Criteria, @entity)
+          dispatch_criteria_to_default_source(
+            :Criteria, data_mapper.entity_class, criteria, data_mapper.attributes
+          )
         end
 
         # @see Method#reader?