+++ spec/unit/repository/query_method_spec.rb
@@ -44,6 +44,13 @@ module Datamappify::Repository::QueryMethod
       it { should_not be_a_writer }
     end
 
+    describe Criteria do
+      subject { Criteria }
+
+      it { should     be_a_reader }
+      it { should_not be_a_writer }
+    end
+
     describe Save do
       subject { Save }